Improve module code generation
Had to do a significant rework of how module identifiers are tokenized,
and how they are parsed, and how they are stored for inference before
I could tackle the improved codegen.

I decided for making it impossible to access partial module paths which
simplified the rest quite a bit. For example if you import
`User.Id` it is now not possible to reference an identifier `User`, you
have to refer to the whole module name with dots.