Remove modules as records because of ambiguities with modules and capitalized identifiers
For example, A.B.C is ambiguous, is this the module A.B.C or is it the
module A.B accessing the identifier C?

This comes up now that I'm going to start adding capitalized identifiers
for the constructors of types.

So for now, I'm going the Elm way to avoid unwanted complexity at this
stage.