Parse type_params in type_constructor with a many helper
And update many of the functions to not have to return the last
position, which actually fixes a small bug with parenthesized types. The
end is now found from the last token processed instead where needed.

And also made an error message for top level definitions consistent with
one a few lines above.